pyproject.toml
[tool.poetry]
name = "leads-analyser"
version = "0.1.0"
description = "Leads Analyser Application"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "apps", from = "." },
]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
whitenoise = "^6.6.0"
drf-spectacular = "^0.27.1"
celery = "^5.3.6"
redis = "^5.0.2"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
